Updated Diamond initialization logic to be part of the contract constructor
- This allows to not have any extra action to be done after the contract is deployed during genesis
- Fixed the tests accordingly
- Updated genesis transaction generation flow accordingly

+++ solidity/supra_contracts/src/SupraContractsBindings.sol
@@ -18,7 +18,7 @@ interface SupraContractsBindings {
     function getCycleStateDetails() external view returns (LibCommon.CycleDetails memory);
 
     // Entry function to be called by node runtime for bookkeeping
-    function processTasks(uint64 _cycleIndex, uint64[] memory _taskIndexes) external;
+    function processTasks(uint64 _cycleIndex, uint256[] memory _taskIndexes) external;
 
     // Entry function of the BlockMeta for block metadata transaction
     function blockPrologue() external;
